+++ train.go
@@ -31,9 +31,9 @@ type TrainOpts struct {
 	// ValidateEvery indicates the number of epochs to run before running a validation. Defaults 1 (every epoch)
 	ValidateEvery int
 
-	CostObserver       func(epoch int, totalEpoch, batch int, totalBatch int, cost float32)
-	ValidationObserver func(confMat ConfusionMatrix, cost float32)
-	MatchTypeFor       func(predVal, targetVal []float32) MatchType
+	CostObserver       func(epoch int, totalEpoch, batch int, totalBatch int, cost float64)
+	ValidationObserver func(confMat ConfusionMatrix, cost float64)
+	MatchTypeFor       func(predVal, targetVal []float64) MatchType
 	CostFn             func(output *gorgonia.Node, accumLoss *gorgonia.Node, target *gorgonia.Node) *gorgonia.Node
 }
 
@@ -90,8 +90,8 @@ func Train(m *Model, layer Layer, trainX, trainY, validateX, validateY tensor.Te
 
 	xShape := append(tensor.Shape{opts.BatchSize}, trainX.Shape()[1:]...)
 
-	x := gorgonia.NewTensor(m.g, tensor.Float32, trainX.Shape().Dims(), gorgonia.WithShape(xShape...), gorgonia.WithName("x"))
-	y := gorgonia.NewMatrix(m.g, tensor.Float32, gorgonia.WithShape(opts.BatchSize, trainY.Shape()[1]), gorgonia.WithName("y"))
+	x := gorgonia.NewTensor(m.g, tensor.Float64, trainX.Shape().Dims(), gorgonia.WithShape(xShape...), gorgonia.WithName("x"))
+	y := gorgonia.NewMatrix(m.g, tensor.Float64, gorgonia.WithShape(opts.BatchSize, trainY.Shape()[1]), gorgonia.WithName("y"))
 
 	result, err := layer(x)
 	if err != nil {
@@ -170,7 +170,7 @@ func Train(m *Model, layer Layer, trainX, trainY, validateX, validateY tensor.Te
 			}
 
 			if opts.CostObserver != nil {
-				opts.CostObserver(i, opts.Epochs, dl.CurrentBatch, dl.Batches, costVal.Data().(float32))
+				opts.CostObserver(i, opts.Epochs, dl.CurrentBatch, dl.Batches, costVal.Data().(float64))
 			} else {
 				// color.Yellow(" Epoch %d %d | cost %v (%v)\n", i, dl.CurrentBatch, costVal, time.Since(startTime))
 			}
